# interactive-video-api/api/storage/docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "Laravel Generator APIs",
        "version": "1.0.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/respostas": {
            "get": {
                "tags": [
                    "Resposta"
                ],
                "summary": "Get a listing of the Respostas.",
                "description": "Get all Respostas",
                "produces": [
                    "application/json"
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/definitions/Resposta"
                                    }
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Resposta"
                ],
                "summary": "Store a newly created Resposta in storage",
                "description": "Store Resposta",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Resposta that should be stored",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Resposta"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Resposta"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "/respostas/{id}": {
            "get": {
                "tags": [
                    "Resposta"
                ],
                "summary": "Display the specified Resposta",
                "description": "Get Resposta",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Resposta",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Resposta"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Resposta"
                ],
                "summary": "Update the specified Resposta in storage",
                "description": "Update Resposta",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Resposta",
                        "required": true,
                        "type": "integer"
                    },
                    {
                        "name": "body",
                        "in": "body",
                        "description": "Resposta that should be updated",
                        "required": false,
                        "schema": {
                            "$ref": "#/definitions/Resposta"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "$ref": "#/definitions/Resposta"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "Resposta"
                ],
                "summary": "Remove the specified Resposta from storage",
                "description": "Delete Resposta",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id of Resposta",
                        "required": true,
                        "type": "integer"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "properties": {
                                "success": {
                                    "type": "boolean"
                                },
                                "data": {
                                    "type": "string"
                                },
                                "message": {
                                    "type": "string"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Resposta": {
            "required": [
                "id_perguntas",
                "descricao_resposta",
                "correta",
                "data_cadastro"
            ],
            "properties": {
                "id_respostas": {
                    "description": "Chave primaria da tabela de respostas",
                    "type": "integer",
                    "format": "int32"
                },
                "id_perguntas": {
                    "description": "Chave estrangeira para tabela de perguntas, uma pergunta tem varias respostas",
                    "type": "integer",
                    "format": "int32"
                },
                "descricao_resposta": {
                    "description": "Descri\u00e7\u00e3o da resposta",
                    "type": "string"
                },
                "correta": {
                    "description": "Resposta correta = 1 errada = 0",
                    "type": "boolean"
                },
                "ativo": {
                    "description": "Status da reposta 1 = ativo 0 = inativo ",
                    "type": "boolean"
                },
                "data_cadastro": {
                    "description": "Data de cadastro da resposta",
                    "type": "string",
                    "format": "date-time"
                },
                "data_atualizacao": {
                    "description": "Data de atualiza\u00e7ao da resposta, atualizado pela aplica\u00e7ao sempre que a resposta for alterada",
                    "type": "string",
                    "format": "date-time"
                }
            }
        },
...
    }
}